(BSR)[PRO] refactor: remove barrel files
