--- pro/src/screens/CollectiveOfferConfirmation/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveOfferConfirmation'
