--- pro/src/screens/CollectiveOfferSummaryCreation/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveOfferSummaryCreation'
