--- pro/src/screens/CollectiveOfferSummaryEdition/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveOfferSummaryEdition'
