--- pro/src/screens/CollectiveOfferVisibility/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollectiveOfferVisibility'
