--- pro/src/screens/CsvTable/index.ts
@@ -1,2 +1,0 @@
-export { default as CsvTableScreen } from './CsvTable'
-export type { TableData as ITableData } from './types'
