--- pro/src/screens/EmailChangeValidation/index.ts
@@ -1,1 +1,0 @@
-export { default as EmailChangeValidationScreen } from './EmailChangeValidation'
