+++ pro/src/screens/IndividualOffer/StocksEventEdition/StockFormList/utils/__specs__/setFormReadOnlyFields.spec.tsx
@@ -1,6 +1,6 @@
 import { OfferStatus } from 'apiClient/v1'
 
-import setFormReadOnlyFields from '../setFormReadOnlyFields'
+import { setFormReadOnlyFields } from '../setFormReadOnlyFields'
 
 describe('setFormReadOnlyFields', () => {
   it('should return empty array when offer is draft', () => {
