--- pro/src/screens/IndividualOffer/StocksEventEdition/StockFormList/utils/index.ts
@@ -1,2 +1,0 @@
-export { buildInitialValues } from './buildInitialValues'
-export { default as setFormReadOnlyFields } from './setFormReadOnlyFields'
